= SipPlatform =  * 64bit 빌드시 warning 을 제거함

+++ trunk/SipPlatform/Log.cpp
@@ -377,8 +377,8 @@ int CLog::GetLogIndex()
 
 static bool LogFileCompare( const std::string & strFirst, const std::string & strSecond )
 {
-	int iFirstLen = strFirst.length();
-	int iSecondLen = strSecond.length();
+	int iFirstLen = (int)strFirst.length();
+	int iSecondLen = (int)strSecond.length();
 
 	int n = strncmp( strFirst.c_str(), strSecond.c_str(), 8 );
